updated chrome extension to build dist that works exactly like test-extension & removed test-extension
