+++ apps/extension/manifest.json
@@ -1,29 +1,20 @@
 {
   "manifest_version": 3,
-  "name": "Primer Extension",
-  "version": "1.0.0",
-  "description": "A Chrome Extension built with React and TypeScript",
-  "permissions": [
-    "activeTab",
-    "storage",
-    "tabs",
-    "scripting"
-  ],
-  "host_permissions": [
-    "https://www.instagram.com/*"
-  ],
-  "action": {
-    "default_popup": "popup.html",
-    "default_title": "Primer Extension"
-  },
+  "name": "Instagram Reels Test Extension",
+  "version": "1.0",
+  "description": "Testing Instagram reels integration for Primer 2.0",
+  "permissions": ["contextMenus", "activeTab", "storage", "scripting", "notifications"],
+  "host_permissions": ["https://www.instagram.com/*"],
   "background": {
     "service_worker": "background.bundle.js"
   },
-  "options_page": "options.html",
-  "content_scripts": [
-    {
-      "matches": ["<all_urls>"],
-      "js": ["content.bundle.js"]
-    }
-  ]
+  "content_scripts": [{
+    "matches": ["https://www.instagram.com/*"],
+    "js": ["content.bundle.js"],
+    "run_at": "document_end"
+  }],
+  "action": {
+    "default_popup": "popup.html",
+    "default_title": "Test Instagram Reels"
+  }
 }
