+++ apps/extension/src/pages/Popup/popup.css
@@ -1,95 +1,70 @@
-.popup-container {
-  width: 300px;
-  height: 400px;
-  padding: 20px;
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  color: white;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif;
+body {
+    width: 350px;
+    padding: 20px;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
+    background: #f5f5dc;
+    color: #333;
+    margin: 0;
 }
 
 .header {
-  text-align: center;
-  margin-bottom: 20px;
+    text-align: center;
+    margin-bottom: 20px;
 }
 
 .header h1 {
-  margin: 0;
-  font-size: 18px;
-  font-weight: 600;
+    margin: 0;
+    font-size: 18px;
+    font-weight: 600;
 }
 
-.content {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
+.status {
+    background: rgba(255, 255, 255, 0.8);
+    padding: 15px;
+    border-radius: 8px;
+    margin-bottom: 15px;
+    border: 1px solid #ddd;
 }
 
-.counter-section {
-  text-align: center;
-  background: rgba(255, 255, 255, 0.1);
-  padding: 20px;
-  border-radius: 8px;
-  backdrop-filter: blur(10px);
+.status.success {
+    background: rgba(76, 175, 80, 0.1);
+    border: 1px solid #4CAF50;
 }
 
-.counter-section h2 {
-  margin: 0 0 15px 0;
-  font-size: 24px;
-  font-weight: 700;
+.status.error {
+    background: rgba(244, 67, 54, 0.1);
+    border: 1px solid #f44336;
 }
 
-.buttons {
-  display: flex;
-  gap: 10px;
-  justify-content: center;
+.image-preview {
+    max-width: 100%;
+    border-radius: 8px;
+    margin: 10px 0;
 }
 
-.btn {
-  padding: 8px 16px;
-  border: none;
-  border-radius: 6px;
-  cursor: pointer;
-  font-size: 14px;
-  font-weight: 500;
-  transition: all 0.2s ease;
+.info {
+    font-size: 12px;
+    opacity: 0.8;
+    margin-top: 10px;
 }
 
-.btn-primary {
-  background: #4CAF50;
-  color: white;
+.button {
+    background: #8b7355;
+    color: white;
+    border: none;
+    padding: 10px 20px;
+    border-radius: 6px;
+    cursor: pointer;
+    width: 100%;
+    font-size: 14px;
+    margin: 5px 0;
 }
 
-.btn-primary:hover {
-  background: #45a049;
-  transform: translateY(-1px);
+.button:hover {
+    background: #7a6248;
 }
 
-.btn-secondary {
-  background: #f44336;
-  color: white;
-}
-
-.btn-secondary:hover {
-  background: #da190b;
-  transform: translateY(-1px);
-}
-
-.tab-info {
-  background: rgba(255, 255, 255, 0.1);
-  padding: 15px;
-  border-radius: 8px;
-  backdrop-filter: blur(10px);
-}
-
-.tab-info h3 {
-  margin: 0 0 10px 0;
-  font-size: 14px;
-  font-weight: 600;
-}
-
-.url {
-  margin: 0;
-  font-size: 12px;
-  word-break: break-all;
-  opacity: 0.8;
+.button:disabled {
+    background: #999;
+    cursor: not-allowed;
 }
