+++ apps/extension/src/pages/Popup/popup.html
@@ -1,53 +1,99 @@
 <!DOCTYPE html>
 <html>
 <head>
-  <meta charset="utf-8">
-  <style>
-    body {
-      width: 300px;
-      height: 400px;
-      margin: 0;
-      padding: 20px;
-      background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-      color: white;
-      font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif;
-    }
-    h1 {
-      margin: 0 0 20px 0;
-      font-size: 18px;
-      font-weight: 600;
-      text-align: center;
-    }
-    .content {
-      background: rgba(255, 255, 255, 0.1);
-      padding: 20px;
-      border-radius: 8px;
-      text-align: center;
-    }
-    .status {
-      background: rgba(76, 175, 80, 0.2);
-      border: 1px solid #4CAF50;
-      padding: 10px;
-      border-radius: 6px;
-      margin: 10px 0;
-    }
-    .info {
-      font-size: 12px;
-      opacity: 0.8;
-      margin-top: 15px;
-    }
-  </style>
+    <meta charset="utf-8">
+    <title>Instagram Reels Test</title>
+    <style>
+        body {
+            width: 350px;
+            padding: 20px;
+            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
+            background: #f5f5dc;
+            color: #333;
+            margin: 0;
+        }
+        
+        .header {
+            text-align: center;
+            margin-bottom: 20px;
+        }
+        
+        .header h1 {
+            margin: 0;
+            font-size: 18px;
+            font-weight: 600;
+        }
+        
+        .status {
+            background: rgba(255, 255, 255, 0.8);
+            padding: 15px;
+            border-radius: 8px;
+            margin-bottom: 15px;
+            border: 1px solid #ddd;
+        }
+        
+        .status.success {
+            background: rgba(76, 175, 80, 0.1);
+            border: 1px solid #4CAF50;
+        }
+        
+        .status.error {
+            background: rgba(244, 67, 54, 0.1);
+            border: 1px solid #f44336;
+        }
+        
+        .image-preview {
+            max-width: 100%;
+            border-radius: 8px;
+            margin: 10px 0;
+        }
+        
+        .info {
+            font-size: 12px;
+            opacity: 0.8;
+            margin-top: 10px;
+        }
+        
+        .button {
+            background: #8b7355;
+            color: white;
+            border: none;
+            padding: 10px 20px;
+            border-radius: 6px;
+            cursor: pointer;
+            width: 100%;
+            font-size: 14px;
+            margin: 5px 0;
+        }
+        
+        .button:hover {
+            background: #7a6248;
+        }
+        
+        .button:disabled {
+            background: #999;
+            cursor: not-allowed;
+        }
+    </style>
 </head>
 <body>
-  <h1>Primer</h1>
-  <div class="content">
-    <h2>Hello World!</h2>
-    <div class="status">
-      ✅ Extension is working!
+    <div class="header">
+        <h1>Instagram Reels Test</h1>
     </div>
-    <p class="info">
-      This is a simple Chrome extension built with React and TypeScript.
-    </p>
-  </div>
+    
+    <div id="status" class="status">
+        <p>Extension loaded. Right-click on an Instagram reel to test!</p>
+    </div>
+    
+    <div id="imageContainer" style="display: none;">
+        <h3>Captured Frame:</h3>
+        <img id="capturedImage" class="image-preview" />
+        <div id="imageInfo" class="info"></div>
+    </div>
+    
+    <button id="testButton" class="button">Test Frame Capture</button>
+    <button id="clearButton" class="button">Clear Results</button>
+    
+    <script src="popup.bundle.js"></script>
 </body>
 </html>
