--- test-extension/manifest.json
@@ -1,20 +1,0 @@
-{
-    "manifest_version": 3,
-    "name": "Instagram Reels Test Extension",
-    "version": "1.0",
-    "description": "Testing Instagram reels integration for Primer 2.0",
-    "permissions": ["contextMenus", "activeTab", "storage", "scripting", "notifications"],
-    "host_permissions": ["https://www.instagram.com/*"],
-    "background": {
-        "service_worker": "background.js"
-    },
-    "content_scripts": [{
-        "matches": ["https://www.instagram.com/*"],
-        "js": ["content.js"],
-        "run_at": "document_end"
-    }],
-    "action": {
-        "default_popup": "popup.html",
-        "default_title": "Test Instagram Reels"
-    }
-}
